+++ src/index.ts
@@ -21,6 +21,7 @@ export { default as DatePicker } from '@/package/components/datepicker.vue';
 export { default as DateRangePicker } from '@/package/components/date-range-picker.vue';
 export { default as TimePicker } from '@/package/components/timepicker.vue';
 export { default as DateTimePicker } from '@/package/components/datetimepicker.vue';
+export { default as Calendar } from '@/package/components/calendar.vue';
 export { default as Pagination } from '@/package/components/pagination.vue';
 export { default as Checkbox } from '@/package/components/checkbox.vue';
 export { default as RadioGroup } from '@/package/components/radio-group.vue';